Fetch missions data from the api
